Tried to solve issue
How to update the UI when user logs in/out? We want to call the function getServerSideProps again, but.. when doing so in useEffect, there will be infinite loop because of strict mode.

Even if I disable strict mode, it won't work. Because when logging out, the user is set to null and only THEN the cookies are being removed. So when getServerSideProps runs again, it still has the cookies and it thinks the user is logged in!
So... we can wait a few seconds and then refresh, but that's a very bad UX