A little bit better Tailwind
Because of the backticks, every space / new line between styles (for readability) is actually being rendered to the DOM and can be seen in dev tools.

This causes "inspect" to be a huge mess.
The new format solves it, but it requires adding comments for every line to prevent prettier from making it one long line.